minor: 将资源模型分散到各个模块中
