+++ gcloud/tasktmpl3/apps.py
@@ -17,3 +17,6 @@ from django.apps import AppConfig
 class Tasktmpl3Config(AppConfig):
     name = 'gcloud.tasktmpl3'
     verbose_name = 'GcloudTasktmpl3'
+
+    def ready(self):
+        from permissions import task_template_resource  # noqa
